requirements.txt
Django==5.2.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
rapidfuzz==3.13.0
Requests==2.32.4
